rename new default provider file
